check file write time before copy dragon to tmp folder
Dragon is a out of process tool and a standalone program. It has to be copied to tmp folder to run. Otherwise if it runs from the gitscc's folder, gitscc cannot be update from the extension manager.

In previous releases, the file is copied every time it needs to run. By checking write time, unnecessary copies are avoid. The tool launches faster.